# Conflicts:
#	src/main/java/com/apiround/greenhub/controller/HomeController.java
#	src/main/java/com/apiround/greenhub/controller/RecipeController.java
#	src/main/resources/application.properties
#	src/main/resources/static/js/region-detail.js
#	src/main/resources/templates/region-detail.html

+++ src/main/java/com/apiround/greenhub/entity/item/ProductPriceOption.java
@@ -15,7 +15,7 @@ public class ProductPriceOption {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "option_id")
-    private Long optionId;
+    private Integer optionId;
 
     @Column(name = "product_id", nullable = false)
     private Integer productId;
